% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritst_genre_table.R
\docType{data}
\name{artist_genre_table}
\alias{artist_genre_table}
\title{Listen Local Artists' Genres}
\format{
A data frame with 1247 rows and 4 variables:
\describe{
  \item{spotify_artist_id}{Spotify Artist ID}
  \item{spotify_artist_name}{The name of the artist on Spotify}
  \item{national_identity}{assumed national identity of the artist(s)}
  \item{spotify_genres}{Genres as identified by Spotify.}
}
}
\source{
\url{http://listenlocal.info/}
}
\usage{
artist_genre_table
}
\description{
A dataset containing national identity and genres of the listen local artists.
}
\keyword{datasets}
