% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_artist_nationality}
\alias{get_artist_nationality}
\title{Artist Nationality}
\usage{
get_artist_nationality(artists, target_ids)
}
\arguments{
\item{artists}{A list of artists with artist IDs}
}
\value{
A logical vector.
}
\description{
Artist Nationality
}
\keyword{internal}
