% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recommendations_genre.R
\name{get_artist_recommendations_genre}
\alias{get_artist_recommendations_genre}
\title{Get Recommendations Based On Genre

Get recommendations based on genre.}
\usage{
get_artist_recommendations_genre(artists_by_genre, target_nationality = "sk")
}
\arguments{
\item{target_nationality}{One of \code{'cz'} or \code{'sk'}.}
}
\value{
A character vector or artist IDs.
}
\description{
Get Recommendations Based On Genre

Get recommendations based on genre.
}
