% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_artist_recommendations.R
\name{get_local_artist_recommendations}
\alias{get_local_artist_recommendations}
\title{Get Slovak Artists IDs

Recommend local artists from user_playlist_artist data frame.}
\usage{
get_local_artist_recommendations(user_playlist_artists, n = 5)
}
\arguments{
\item{user_playlist_artists}{A user_playlist_artist data frame.}

\item{n}{Number of required recommendations (maximum value)}
}
\value{
Returns n or less artist ids.
}
\description{
Get Slovak Artists IDs

Recommend local artists from user_playlist_artist data frame.
}
\examples{
get_national_artist_ids ( national_identity = 'sk' )
}
