% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_recommendations.R
\name{get_local_recommendations}
\alias{get_local_recommendations}
\title{Get Local Recommendations}
\usage{
get_local_recommendations(
  user_playlist_id = "6KHw5aZWWsmRqpT7o290Mo",
  target_nationality = "sk",
  target_release = NULL,
  limit = 20,
  n = 4
)
}
\arguments{
\item{target_nationality}{Defaults to \code{"sk"}}

\item{target_release}{If recommendation should be limited to target,
defaults to \code{NULL}}

\item{limit}{Number of playlist items used for recommendation seed.}

\item{n}{number of required target country recommendations}

\item{playlist_information}{A list received from get_playlist_information.}
}
\value{
A tibble of recommendations.
}
\description{
Get Local Recommendations
}
