% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_artist_ids.R
\name{get_national_artist_ids}
\alias{get_national_artist_ids}
\title{Get Slovak Artists IDs

Get unique tracks and unique artists from a playlist in a list.}
\usage{
get_national_artist_ids(national_identity = "sk")
}
\arguments{
\item{national_identity}{One of \code{'cz'} or \code{'sk'}.}
}
\value{
A character vector or artist IDs.
}
\description{
Get Slovak Artists IDs

Get unique tracks and unique artists from a playlist in a list.
}
\examples{
get_national_artist_ids ( national_identity = 'sk' )
}
