% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_tracks.R
\name{get_nearest_tracks}
\alias{get_nearest_tracks}
\title{Get Nearest Tracks

Get recommendations based on genre.}
\usage{
get_nearest_tracks(user_tracks, new_tracks, n = 5)
}
\arguments{
\item{user_tracks}{A tibble created by \code{\link{ get_playlist_information}}.}

\item{new_tracks}{Tracks to limit.}

\item{n}{Number of recommendations needed}
}
\value{
A character vector or artist IDs.
}
\description{
Get Nearest Tracks

Get recommendations based on genre.
}
