% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_country.R
\name{get_release_country}
\alias{get_release_country}
\title{Get Release Country}
\usage{
get_release_country(isrc)
}
\arguments{
\item{isrc}{A vector of iscr codes}
}
\description{
Get Release Country
}
