% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_recommendations_artists.R
\name{get_track_recommendations_artist}
\alias{get_track_recommendations_artist}
\title{Get recommendations for an artist}
\usage{
get_track_recommendations_artist(
  spotify_artist_id,
  target_nationality = "sk",
  target_release = NULL,
  n = 5
)
}
\arguments{
\item{spotify_artist_id}{A vector of artist IDs}

\item{target_nationality}{Defaults to \code{"sk"}}

\item{target_release}{If recommendation should be limited to target,
defaults to \code{NULL}}

\item{n}{Number of recommended tracks needed.}
}
\value{
A tibble of recommendations.
}
\description{
Get recommendations for an artist
}
