% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listen_local_artists.R
\docType{data}
\name{listen_local_artists}
\alias{listen_local_artists}
\title{Listen Local Artists}
\format{
A data frame with 864 rows and 4 variables:
\describe{
  \item{spotify_artist_id}{Spotify Artist ID}
  \item{artist_name}{The name of the artist on Spotify}
  \item{national_identity}{assumed national identity of the artist(s)}
  \item{language}{language of recordings}
}
}
\source{
\url{http://listenlocal.info/}
}
\usage{
listen_local_artists
}
\description{
A dataset containing national identity, language and cities of
artists.
}
\keyword{datasets}
