% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_genre_table.R
\docType{data}
\name{local_genre_table}
\alias{local_genre_table}
\title{Local Genre Table}
\format{
A data frame with 463281 rows and 3 variables:
\describe{
  \item{genre}{A simplified genre description.}
  \item{genre_rec}{A long genre description to be simplified.}
  \item{distance}{A numeric distance}
}
}
\source{
\url{https://reprex.nl/project/listen-local/}
}
\usage{
local_genre_table
}
\description{
A dataset containing a vocabulary table for English language genre description simplifications.
}
\keyword{datasets}
