% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_recommendations.R, R/params.R
\name{get_local_recommendations}
\alias{get_local_recommendations}
\alias{recparam}
\title{Get Local Recommendations}
\usage{
get_local_recommendations(
  playlist_id = "6KHw5aZWWsmRqpT7o290Mo",
  target_nationality = "sk",
  target_release = "sk",
  recommendation_type = "artists",
  limit = 20,
  n_rec = 4,
  silent = TRUE,
  authorization = NULL,
  listen_local_artist_data = "not_included/listen_local_artists.rds",
  artist_distance_data = "not_included/artist_distances.rds",
  local_genre_table_path = "not_included/local_genre_table.rds",
  artist_genre_table_path = "not_included/artist_genre_table.rds"
)
}
\arguments{
\item{playlist_id}{A Spotify playlist id. This playlist serves as the basis of the recommendation.}

\item{target_nationality}{Defaults to \code{"sk"}. The nationality id of the recommendation candidate artists, if necessary.}

\item{target_release}{Defaults to \code{NULL}. The country of origin of the recommendation candidate tracks, if necessary.}

\item{recommendation_type}{Defaults to \code{'artists'}, can be \code{'release'}
or \code{'both'}.}

\item{limit}{Number of playlist items used for recommendation seed.}

\item{n_rec}{Number of required recommendations. (Maximum value, may not be fulfilled.)}

\item{silent}{Defaults to \code{TRUE}.}

\item{authorization}{Defaults to \code{NULL} when
\code{\link[spotifyr]{get_spotify_access_token}} is invoked.}

\item{listen_local_artist_data}{If \code{NULL}, the functions try to load the data from the
spotifyrecommendations package.}

\item{artist_distance_data}{If \code{NULL}, the functions try to load the data from the
spotifyrecommendations package.}

\item{local_genre_table_path}{If \code{NULL}, the functions try to load the data from the
spotifyrecommendations package.}

\item{artist_genre_table_path}{If \code{NULL}, the functions try to load the data from the
spotifyrecommendations package.}
}
\value{
A tibble of recommendations.
}
\description{
This is the wrapper function to the Listen Local Demo recommendation system.
}
