% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_of_dutch_indie_playlist.R
\docType{data}
\name{sound_of_dutch_indie_playlist}
\alias{sound_of_dutch_indie_playlist}
\title{The Sound of Dutch Indie Playlist}
\format{
A nested list.
\describe{
  \item{collaborative}{A boolean}
  \item{description}{A text description of the playlist}
  \item{external_urls}{Distance between base artist and recommended artist}
  \item{followers}{assumed national identity of the artist(s)}
  \item{href}{https://api.spotify.com/v1/playlists/2pSeH1YQCSICs2knjs7e5o}
  \item{id}{2pSeH1YQCSICs2knjs7e5o}
  \item{images}{Image URL and dimensions, a list.}
  \item{name}{The Sound of Dutch Indie}
  \item{owner}{thesoundsofspotify}
  \item{primary_color}{Primary color of the list}
  \item{public}{A boolean}
  \item{snapshot_id}{An ID}
  \item{tracks}{A list of tracks.}
  \item{type}{Object type - playlist.}
  \item{uri}{spotify:playlist:2pSeH1YQCSICs2knjs7e5o}
}
}
\source{
\url{https://open.spotify.com/playlist/2pSeH1YQCSICs2knjs7e5o}
}
\usage{
sound_of_dutch_indie_playlist
}
\description{
A dataset containing \code{The Sound of Dutch Indie} playlist for local testing.
}
\keyword{datasets}
