% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{spotify_artist_id}
\alias{spotify_artist_id}
\title{Artist's Unique ID on Spotify}
\arguments{
\item{spotify_artist_id}{Artist's Unique ID on Spotify}
}
\description{
An analysis of a user playlist that will serve as the basis of new recommendations.
}
