% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_artist_recommendations.R,
%   R/initial_recommendations.R, R/params.R
\name{get_local_artist_recommendations}
\alias{get_local_artist_recommendations}
\alias{initial_recommendations}
\alias{target_ids}
\title{Get Slovak Artists IDs}
\usage{
get_local_artist_recommendations(
  user_playlist_artists,
  target_ids,
  n_rec = 5,
  authorization = NULL
)

initial_recommendations(
  playlist_information,
  target_ids = NA_character_,
  limit = 20,
  silent = TRUE,
  authorization = NULL
)
}
\arguments{
\item{user_playlist_artists}{A user_playlist_artists data frame from
\code{\link{get_playlist_information}}.}

\item{target_ids}{A character vector of the suitable recommendation candidates.}

\item{n_rec}{Number of required recommendations. (Maximum value, may not be fulfilled.)}

\item{authorization}{Defaults to \code{NULL} when
\code{\link[spotifyr]{get_spotify_access_token}} is invoked.}

\item{playlist_information}{A list received from A playlist queried by
\code{{get_playlist_information}}.}

\item{limit}{How many items are used for the initial recommendation.}

\item{silent}{Defaults to \code{TRUE}.}
}
\value{
Returns a character vector of n or less artist ids.

A tibble of recommendations. \code{recommendations_by_tracks$target_artists} contains a
logical vector, identifying the tracks that are among \code{target_ids}.
}
\description{
Recommend local artists from user_playlist_artist data frame.
}
