% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_playlist.R
\name{upload_playlist}
\alias{upload_playlist}
\title{Upload Playlist

Get unique tracks and unique artists from a playlist in a list.}
\usage{
upload_playlist(
  user_id = "rx4xjay1368opqg2i7nabuo5c",
  playlist_name = "test",
  playlist_description = "listen local test"
)
}
\arguments{
\item{user_id}{A single Spotify playlist id}

\item{playlist_name}{Name for a new playlist}

\item{playlist_description}{New description for the playlist}
}
\value{
Does not return anything.
}
\description{
Upload Playlist

Get unique tracks and unique artists from a playlist in a list.
}
