% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{user_playlist_artists}
\alias{user_playlist_artists}
\title{Information about the artists on the user playlist}
\arguments{
\item{user_playlist_artists}{A user_playlist_artists data frame from
\code{\link{user_playlist_info}}.}
}
\description{
An analysis of a user playlist that will serve as the basis of new recommendations.
}
