% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{user_playlist_info}
\alias{user_playlist_info}
\title{Information about the user playlist}
\arguments{
\item{user_playlist_info}{An analysis of a user playlist that will serve as the basis of new recommendations
created by \code{get_playlist_information}.}
}
\description{
An analysis of a user playlist that will serve as the basis of new recommendations.
}
