% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_genre.R
\name{get_artist_genre}
\alias{get_artist_genre}
\title{Get Artist Genre

Get the genres of the artist, if available.}
\usage{
get_artist_genre(user_playlist_artists)
}
\arguments{
\item{user_playlist_artists}{A user_playlist_artists data frame from
\code{\link{user_playlist_info}}}
}
\value{
A tibble of artist IDs and genres in long form.
}
\description{
Get Artist Genre

Get the genres of the artist, if available.
}
