% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_recommendations.R
\name{get_local_recommendations}
\alias{get_local_recommendations}
\title{Get Local Recommendations}
\usage{
get_local_recommendations(
  user_playlist_id = "6KHw5aZWWsmRqpT7o290Mo",
  target_country = "sk",
  recommendation_type = "artists",
  limit = 20,
  n = 4,
  authorization = NULL
)
}
\arguments{
\item{target_country}{Defaults to \code{"sk"}}

\item{limit}{Number of playlist items used for recommendation seed.}

\item{n}{number of required target country recommendations}

\item{authorization}{Defaults to \code{NULL} when
\code{get_spotify_access_token()} is invoked.}

\item{playlist_information}{A list received from get_playlist_information.}
}
\value{
A tibble of recommendations.
}
\description{
Get Local Recommendations
}
