% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_artist_ids.R
\name{get_national_artist_ids}
\alias{get_national_artist_ids}
\title{Get Artists IDs For Target Nationality.}
\usage{
get_national_artist_ids(target_identity = "sk")
}
\arguments{
\item{target_identity}{One of \code{'sk'} or \code{'nl'} or \code{'hu'}.}
}
\value{
A character vector or artist IDs.
}
\description{
Get unique tracks and unique artists from a playlist in a list.
}
\examples{
get_national_artist_ids ( national_identity = 'sk' )
}
