% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_information.R
\name{get_playlist_information}
\alias{get_playlist_information}
\title{Get Playlist Information}
\usage{
get_playlist_information(playlist_id)
}
\arguments{
\item{playlist_id}{A single Spotify playlist id}
}
\value{
A list of unique tracks and unique artists from a playlist.
}
\description{
Get Playlist Information
}
