% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_country.R
\name{get_release_country}
\alias{get_release_country}
\title{Get Release Country

Get the country codes of the release countries by ISRC code.}
\usage{
get_release_country(isrc)
}
\arguments{
\item{isrc}{A vector of iscr codes}
}
\value{
A vector of ISO-3166 country codes.
}
\description{
Get Release Country

Get the country codes of the release countries by ISRC code.
}
