% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_recommendations.R
\name{initial_recommendations}
\alias{initial_recommendations}
\title{Get Initial Recommendations

The initial recommendations are based on the Spotify algorithm, and try to
meet the user local content targets.}
\usage{
initial_recommendations(
  playlist_information,
  target_ids = NA_character_,
  limit = 20,
  silent = TRUE,
  authorization = NULL
)
}
\arguments{
\item{playlist_information}{A list received from A playlist queried by
\code{{get_playlist_information}}.}

\item{target_ids}{Spotify artists ids from the targeted repertoire.}

\item{limit}{How many items are used for the initial recommendation.}

\item{authorization}{Defaults to \code{NULL} when #' \code{get_spotify_access_token()} is invoked.}

\item{target_release}{The targetted country code of the sound recordings,}
}
\value{
A tibble of recommendations. \code{recommendations_by_tracks$target_artists} contains a
logical vector, identifying the tracks that are among \code{target_ids}.
}
\description{
Get Initial Recommendations

The initial recommendations are based on the Spotify algorithm, and try to
meet the user local content targets.
}
