% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listen_local_artists.R
\docType{data}
\name{listen_local_artists}
\alias{listen_local_artists}
\title{Listen Local Artists}
\format{
A data frame with 1110 rows and 3 variables:
\describe{
  \item{spotify_artist_id}{Spotify Artist ID}
  \item{spotify_artist_name}{The name of the artist on Spotify}
  \item{national_identity}{assumed national identity of the artist(s)}
}
}
\source{
\url{https://reprex.nl/project/listen-local/}
}
\usage{
listen_local_artists
}
\description{
A dataset containing national identity of listen local artists.
}
\keyword{datasets}
