% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_playlist.R
\name{upload_playlist}
\alias{upload_playlist}
\title{Upload Playlist}
\usage{
upload_playlist(
  user_id = NULL,
  playlist_name = "test_listen_local_playlist",
  playlist_description = "A test playlist of the Listen Local App",
  track_uris = NULL,
  authorization = NULL
)
}
\arguments{
\item{user_id}{A single Spotify playlist id}

\item{playlist_name}{Name for a new playlist}

\item{playlist_description}{New description for the playlist}

\item{track_uris}{The track uris of the new playlist.}

\item{authorization}{Defaults to \code{NULL} when
\code{\link[spotifyr]{get_spotify_access_token}} is invoked.}
}
\value{
Does not return anything.
}
\description{
Upload the recommended playlist to the user's Spotify profile.
}
